Add modules folder, __mock__ and test folders
